decimalSeparator/thousandSeparator code improvements
