+++ src/main/java/org/primefaces/util/ComponentUtils.java
@@ -559,14 +559,17 @@ public class ComponentUtils {
 
     /**
      * Tries to retrieve value from stateHelper by key first. If the value is not present (or is null),
-     * then it is retrieved from defaultValueSupplier
+     * then it is retrieved from defaultValueSupplier.
+     *
+     * Should be removed when {@link StateHelper} is extended with similar functionality.
+     * (see https://github.com/eclipse-ee4j/mojarra/issues/4568 for details)
      * @param stateHelper The stateHelper to try to retrieve value from
      * @param key The key under which value is stored in the stateHelper
      * @param defaultValueSupplier The object, from which default value is retrieved
      * @param <T> the expected type of returned value
      * @return value from stateHelper or defaultValueSupplier
      */
-    public static <T> T getValue(StateHelper stateHelper, Serializable key, Supplier<T> defaultValueSupplier) {
+    public static <T> T eval(StateHelper stateHelper, Serializable key, Supplier<T> defaultValueSupplier) {
         T value = (T) stateHelper.eval(key, null);
         if (value == null) {
             value = defaultValueSupplier.get();
